requirements.txt
dnspython
dotenv
python-dotenv
